unbucketize

+++ import_datafile_to_database.py
@@ -4,6 +4,8 @@ import sys
 import os
 
 path = sys.argv[1]
+bucketize = False
+
 if "label" in path:
     with_label = True
 else:
@@ -28,11 +30,12 @@ try:
     else:
         drop_and_create_table(db_conn, "darpa_sample", "source varchar(20), destination varchar(20), timestamp varchar(20)")
         exec_sql(db_conn, "copy darpa_sample(source, destination, timestamp) from '%s' delimiter ','" % path)
-
-    exec_sql(db_conn, "alter table darpa_sample alter column timestamp type varchar(13) using SUBSTR(timestamp, 1, 13)")
-    drop_and_copy_table(db_conn, "darpa_sample", "darpa_sample_distinct", "distinct on (source, destination, timestamp) *")
-    drop_table(db_conn, "darpa_sample")
-    exec_sql(db_conn, "alter table darpa_sample_distinct rename to darpa_sample")
+    
+    if bucketize:
+        exec_sql(db_conn, "alter table darpa_sample alter column timestamp type varchar(13) using SUBSTR(timestamp, 1, 13)")
+        drop_and_copy_table(db_conn, "darpa_sample", "darpa_sample_distinct", "distinct on (source, destination, timestamp) *")
+        drop_table(db_conn, "darpa_sample")
+        exec_sql(db_conn, "alter table darpa_sample_distinct rename to darpa_sample")
     print "Data copied to table darpa_sample"
 except KeyboardInterrupt:
     print "exit manually"
